backend/requirements.txt
# Web框架
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.11.1
pydantic-settings==2.1.0
email-validator==2.2.0

# 数据库
neo4j==5.15.0
sqlalchemy==2.0.27
alembic==1.13.1

# 自然语言处理
bertopic==0.15.0
sentence-transformers==2.5.1
jieba==0.42.1
stanza==1.5.0
transformers==4.37.2
torch==2.2.0
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
umap-learn==0.5.5
hdbscan==0.8.33
spacy==3.8.5
gensim==4.3.3

# LLM和向量数据库
llama-index==0.12.28
openai==1.70.0
tiktoken==0.9.0

# 工具包
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.2
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
PyJWT==2.10.1 